% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{clean.otu}
\alias{clean.otu}
\title{Clean OTU table from empty rows}
\usage{
clean.otu(otu, tax)
}
\arguments{
\item{otu}{the input dataframe (eDNA reads) which the rows will be removed}

\item{tax}{the input dataframe (taxonomic information) which the rows will be removed}
}
\value{
OTU table without rows that contains 0's across all stations
}
\description{
This function removes the empty rows from the OTU table and is also removes them from
the TAX dataframe. This functions automatically updates the otu and tax table as it creates
2 objects named otu and tax.
It will print out which tax was removed
}
\examples{
clean.otu(otu,tax)
taxa_removed <- clean.otu(otu,tax)
}
