% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{coll_sim2}
\alias{coll_sim2}
\title{Collapses dataframe 2}
\usage{
coll_sim2(db, sim)
}
\arguments{
\item{db}{the database needed to be collapsed}

\item{sim}{The vector which similarities are taken from}
}
\value{
A database with collapsed rows (Fun = sum) based on the vector
}
\description{
This function Collapses a dataframe based on a similar vector
}
\examples{
otu <- coll_sim2(edna_to_otu(edna),tax_to_vector(comb_tax_compared))
}
