% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{collapse_meta}
\alias{collapse_meta}
\title{Collapse metadata table based on chosen level}
\usage{
collapse_meta(metadata)
}
\arguments{
\item{metadata}{metadata where all stations are rows}
}
\value{
collapsed metadata table based on chosen level
}
\description{
This function for collapsing otu statoins (columns) based on n=column name in metadata
}
\examples{
mm<-collapse_meta #remember to collapse on the column "collapse.name"
}
