% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{collapse_similar_rows}
\alias{collapse_similar_rows}
\title{Collapse similar rows on a dataframe}
\usage{
collapse_similar_rows(db, sim)
}
\arguments{
\item{db}{The dataframe or vector that should be collapsed}

\item{sim}{The vector indicating similarities}
}
\value{
A dataframe or vector containing the collapsed(unique) valued
}
\description{
This function collapses similar rows or based on an index. If the db is a vector then the vector will
be collapsed to a shorter vector based on the sim vector
}
\examples{
collapse_similar_rows(edna[,st:end],edna$tax_id)
}
