% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{colmax}
\alias{colmax}
\title{Colmax}
\usage{
colmax(input)
}
\arguments{
\item{input}{A dataframe (numeric)}
}
\value{
A vector with column maximums
}
\description{
This function computes the column maximum
}
\examples{
colmin(OTU)
}
