% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{colmin}
\alias{colmin}
\title{Colmin}
\usage{
colmin(input)
}
\arguments{
\item{input}{A dataframe (numeric)}
}
\value{
A vector with column minimums
}
\description{
This function computes the column minimum
}
\examples{
colmin(OTU)
}
