% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{colspectrum}
\alias{colspectrum}
\title{Spectrum colours}
\usage{
colspectrum(n)
}
\arguments{
\item{n}{The number of colours to be produced}
}
\value{
R vector
}
\description{
This function generates n number of colours that are in a spectrum
}
\examples{
colspectrum(5)
[1] "#FF0000" "#FFBF00" "#7FFF3F" "#10D997" "#4169E1"
}
