% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{copy}
\alias{copy}
\title{Copies the rownames and columnames of a data.frame into another one}
\usage{
copy(df)
}
\arguments{
\item{df}{The data.frame that the values will be retreived}
}
\value{
generates a data.frane with the same colnames and rownames
}
\description{
This function creates a pie-chart on the diversity of MOTUs found grouped by the "rank"
}
\examples{
copy(edna)
copy(taxa)
}
