% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{custom_cbind}
\alias{custom_cbind}
\title{Custom cbind Function for Dataframes with Different Rows}
\usage{
custom_cbind(q_1, w_1)
}
\arguments{
\item{q_1}{A dataframe. The first dataframe to be column-bound.}

\item{w_1}{A dataframe. The second dataframe to be column-bound.}
}
\value{
A dataframe that is the result of column-binding \code{q_1} and \code{w_1}.
If the dataframes have different rows, missing rows in each dataframe
will be added with \code{NA} values before binding.
}
\description{
This function combines two dataframes with different rows by adding
the missing rows with \code{NA} values and then performing a column bind
on the resulting dataframes.
}
\examples{
q_1 <- data.frame(A = 1:3, B = letters[1:3])
w_1 <- data.frame(C = 4:5, D = letters[4:5])

custom_cbind(q_1, w_1)
}
