% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{custom_rbind}
\alias{custom_rbind}
\title{Custom rbind Function for Dataframes with Different Columns}
\usage{
custom_rbind(df1, df2)
}
\arguments{
\item{df1}{A dataframe. The first dataframe to be row-bound.}

\item{df2}{A dataframe. The second dataframe to be row-bound.}
}
\value{
A dataframe that is the result of row-binding \code{df1} and \code{df2}.
If the dataframes have different columns, missing columns in each dataframe
will be added with \code{NA} values before binding.
}
\description{
This function combines two dataframes with different columns by adding
the missing columns with \code{NA} values and then performing a row bind
on the resulting dataframes.
}
\examples{
df1 <- data.frame(A = 1:3, B = letters[1:3])
df2 <- data.frame(A = 4:5, C = letters[4:5])

custom_rbind(df1, df2)

}
