% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{dimentions}
\alias{dimentions}
\title{Check of dimentions of otu, mm, and tax}
\usage{
dimentions(otu = otu, mm = mm, tax = tax)
}
\arguments{
\item{otu}{The number of colours to be produced}

\item{mm}{The number of colours to be produced}

\item{tax}{The number of colours to be produced}
}
\value{
message output
}
\description{
This function checks that otu table, metadata (mm) table and the tax table are on the same dimensions.
That means checking if there are the same number of species on the otu table that they are in the tax table.
And it also checks that there are the same number of stations in otu table and metadata (mm)
}
\examples{
dimentions(otu,mm,tax)
}
