% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{inv.trans}
\alias{inv.trans}
\title{Inverse transpose dataframe}
\usage{
inv.trans(input.X, input.Y, input.Z)
}
\arguments{
\item{input.X}{the vector where the columns/colnames are stored}

\item{input.Z}{the vector where the values are stored}
}
\value{
A two dimentional dataframe
}
\description{
This function does the inverse of trans() function. It transposes 1 dimentional dataframe
into a 2 dimentional dataframe
}
\examples{
clean.otu(otu,tax)
taxa_removed <- clean.otu(otu,tax)
}
