% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{inverselogit}
\alias{inverselogit}
\title{Inverse logit function}
\usage{
inverselogit(x)
}
\arguments{
\item{x}{The value in logit scale}
}
\value{
The vale on the normal scale
}
\description{
This function returns the inverse logit value
}
\examples{
inverselogit(0.01)
}
