% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{le}
\alias{le}
\title{Unique length}
\usage{
le(input)
}
\arguments{
\item{input}{A vector}
}
\value{
A number of unique elements in the input vector
}
\description{
This function gives the length of unique elements in a vector
}
\examples{
colmin(OTU)
}
