% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(x)
}
\arguments{
\item{x}{The value in normal scale \link{0:1} values}
}
\value{
The vale on the logit scale
}
\description{
This function returns the logit value. This function transforms values between 0 and 1 into logit-scale values
}
\examples{
logit(0.01)
}
