% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{low}
\alias{low}
\title{Set up a working directory (easily to convert between windows and mac)}
\usage{
low(folder = "", m = F)
}
\arguments{
\item{folder}{The path to the folder (After the path to pd) that you want to set the working directory}

\item{m}{logical value if you want the menu to list all the directories below that folder}
}
\value{
generates a working directory
}
\description{
This function creates a pie-chart on the diversity of MOTUs found grouped by the "rank"
}
\examples{
low(folder="eDNA",m=F)
low(folder="",m=T)

}
