% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{make.unique.2}
\alias{make.unique.2}
\title{Unique duplicated names}
\usage{
make.unique.2(x, sep = ".")
}
\arguments{
\item{x}{the text string containing duplicated values}

\item{sep}{the seperation character}
}
\value{
text string
}
\description{
This function creates unique names for all the duplicated values. It takes into account the first duplicate value/name as well
}
\examples{
make.unique.2(species_names)
make.unique.2(edna$genus_name)
}
