% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{match_multiple}
\alias{match_multiple}
\title{Match multiple inputs (grep in for multiple variables at once)}
\usage{
match_multiple(pattern, x)
}
\arguments{
\item{pattern}{The pattern that that you want to search for. Can be a single variable or a vector}

\item{x}{The variable or the vector that you want to look in}
}
\value{
The matching values
}
\description{
This function serches and matches multiple inputs at once
}
\examples{
match_multiple("Squalius acanthius",dataframe)
match_multiple(c("sp1","sp2"),vector)
}
