% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{motudiv}
\alias{motudiv}
\title{Creates a piechart on MOTU diversity}
\usage{
motudiv(taxa, rank, logic = 1)
}
\arguments{
\item{taxa}{Dataframe that contains taxa (each rank in one column)}

\item{rank}{The "rank" that you are interested in grouping the MOTUs to}

\item{logic}{If logic=2 the "unassigned" MOTUs will be removed. If logic=1 the "unassigned" MOTUs will remain}
}
\value{
Creates a pie-chart with the diversity of MOTUs per "rank"
}
\description{
This function creates a pie-chart on the diversity of MOTUs found grouped by the "rank"
}
\examples{
motudiv(taxa,"species")
motudiv(taxa,"phylum_name")
}
