% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{multi_grep}
\alias{multi_grep}
\title{Multigrep}
\usage{
multi_grep(pattern, x)
}
\arguments{
\item{pattern}{The pattern that that you want to search for. Can be a single variable or a vector}

\item{x}{The variable or the vector that you want to look in}
}
\value{
The matching values
}
\description{
This function does grep but allows for multiple inputs (so a vector)
}
\examples{
match_multiple("Squalius acanthius",dataframe)
match_multiple(c("sp1","sp2"),vector)
}
