% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{pa_conversion}
\alias{pa_conversion}
\title{Collapse metadata table based on chosen level}
\usage{
pa_conversion(otu, threshold = 0)
}
\arguments{
\item{otu}{otu table where species are rows and columns are stations}

\item{threshold}{the threshold where a MOTU is considered "Present"}
}
\value{
otu table with only presence / absence
}
\description{
This function for collapsing otu statoins (columns) based on n=column name in metadata
}
\examples{
otu<-pa_conversion(otu)
}
