% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{param.extract}
\alias{param.extract}
\title{Extract the Stan model parameter names}
\usage{
param.extract(stanMod)
}
\arguments{
\item{stanMod}{The stan model as a S4 (Stan list)}
}
\value{
A vector of all the names from the parameter
}
\description{
This function extracts the names of the parameters from a Stan model output in R
}
\examples{
param.extract(stanMod)
[1] "phi_0" "phi_1" "theta" "lp__" 
}
