% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{prep_ridg_plot}
\alias{prep_ridg_plot}
\title{Prepare the parameters output from stanMod for ridgit plots}
\usage{
prep_ridg_plot(parameter, fa1, fa2, faidx, x = 2, y = 1)
}
\arguments{
\item{parameter}{The name of the parameter that will be used to be extracted}

\item{fa1}{Factor 1}

\item{fa2}{Factor 2}

\item{faidx}{The vector which the parameter has been indexed in the stanModel}

\item{x}{Factor 1 on the y axis}

\item{y}{Factor 2 as different colours}
}
\value{
A dataframe ready to be as the input for geom_density_ridges() in ggplot2
}
\description{
This function conducts data manipulation for transforming the parameters output from stanMod into a format that
required for geom_density_ridges() in ggplot2.
}
\examples{
tparam <- prep_ridg_plot(parameter,fa1,fa2,faidx,2,1)
ggplot(tparam,aes(x = `Z`, y = `X`, fill = `Y`))+#
geom_density_ridges(scale = 2,size = 0.2,rel_min_height = 0.01,alpha = .8)
}
