% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{printbold}
\alias{printbold}
\title{Print bold}
\usage{
printbold(input, wide = 40, col = 47)
}
\arguments{
\item{input}{the text that will be printed}

\item{wide}{How wide should be the print result}

\item{col}{the colur of the text in between ##}
}
\value{
prints on Console
}
\description{
This function prints on Console with ##
}
\examples{
printbold("hehehe")
printbold("hehehe", wide=60)
printbold("hehehe", wide=60, col=38)
}
