% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{questions}
\alias{questions}
\title{Generates string text with colours}
\usage{
questions(input, col = 43)
}
\arguments{
\item{input}{the text string to be printed}

\item{col}{the code for the color}
}
\value{
text string
}
\description{
This function returns a text with colourw
}
\examples{
questions("hehehe")
hehehe (colored)
}
