% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{stan_code_pdf}
\alias{stan_code_pdf}
\title{Stan code into PDF converter}
\usage{
stan_code_pdf(stanMod, title)
}
\arguments{
\item{stanMod}{The stan model output where the compiled stan code is.}

\item{title}{A title for the PDF file}
}
\value{
A PDF file in your working directory getwd() with the title given
}
\description{
This function prints out a PDF file with the stan code. The stan output from running the
stan(file,data) function in 'rstan' package contains the Stan code used for generating the output,
thus that code will be compiled into a PDF file.
}
\examples{
stan_code_pdf(stanMod,"Model_1")
}
