% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{trans}
\alias{trans}
\title{Transposes a data.frame from 2 dimensional to 1 dimensional}
\usage{
trans(df)
}
\arguments{
\item{df}{The data.frame that has the vlaues in the cell as a combination of col and rows}
}
\value{
generates a data.frane with $X = rows, $Y = columns and $Z = the value of the cells
}
\description{
This function creates a pie-chart on the diversity of MOTUs found grouped by the "rank"
}
\examples{
trans(edna)
trans(abundance.data)
}
