% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{v}
\alias{v}
\title{Change the parameters in the plot view}
\usage{
v(x = 1, y = 1)
}
\arguments{
\item{x, y}{input x and y}
}
\value{
the plot window is separated in x rows and y columns
}
\description{
Change the parameters in the plot view
}
\examples{
v(2,2)
}
