% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{inverselogit}
\alias{inverselogit}
\title{Inverse logit function}
\usage{
inverselogit(x)
}
\arguments{
\item{x}{The value in logit scale}
}
\value{
The backtransformed vale on the normal scale \link{0:1}
}
\description{
This function returns the inverse logit value. This function transforms values from logit into values between 0 and 1
}
\examples{
inverselogit(0.01)
}
