% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggu.R
\name{logit}
\alias{logit}
\title{Logit function}
\usage{
logit(x)
}
\arguments{
\item{x}{The value in normal scale}
}
\value{
The vale on the logit scale
}
\description{
This function returns the logit value
}
\examples{
logit(0.01)
}
